naive implementation for host scanning
