+++ modules/hostdb/hostentry.go
@@ -33,47 +33,12 @@ func (hdb *HostDB) priceWeight(entry modules.HostEntry) (weight types.Currency) 
 	return baseWeight.Div(price).Div(price).Div(price)
 }
 
-// threadedProbeHost tries to fetch the settings of a host. If successful, the
-// host is put in the set of active hosts. If unsuccessful, the host id deleted
-// from the set of active hosts.
-func (hdb *HostDB) threadedProbeHost(entry *modules.HostEntry) {
-	// If the host is already in the set of active hosts, remove it. It will be
-	// re-added if it responds to the probing, and left out if not. The db
-	// needs to be locked for this operation, but then unlocked while the
-	// network communication is happening.
-	id := hdb.mu.Lock()
-	{
-		node, exists := hdb.activeHosts[entry.IPAddress]
-		if exists {
-			delete(hdb.activeHosts, entry.IPAddress)
-			node.removeNode()
-		}
-	}
-	hdb.mu.Unlock(id)
-
-	// Request the most recent set of settings from the host.
-	var settings modules.HostSettings
-	err := hdb.gateway.RPC(entry.IPAddress, "HostSettings", func(conn modules.NetConn) error {
-		return conn.ReadObject(&settings, 1024)
-	})
-	if err != nil {
-		return
-	}
-
-	// Re-insert the host into the database as it is online and has responded
-	// with a set of settings.
-	id = hdb.mu.Lock()
-	defer hdb.mu.Unlock(id)
-	entry.HostSettings = settings
-	entry.Weight = hdb.priceWeight(*entry)
-	hdb.insertNode(entry)
-}
-
 // insert adds a host entry to the state. The host will be inserted into the
 // set of all hosts, and if it is online and responding to requests it will be
 // put into the list of active hosts.
 func (hdb *HostDB) insertHost(entry modules.HostEntry) {
 	// Add the host to allHosts.
+	entry.Reliability = types.NewCurrency64(5)
 	hdb.allHosts[entry.IPAddress] = &entry
 
 	// TODO: Add sybil attack prevention mechanism.
